requirements.txt
# Core dependencies
fastapi==0.104.1
pydantic<2.0.0  # PyCaret 3.2.0 is not compatible with pydantic 2.0+
python-dotenv==1.0.0
python-multipart==0.0.6
python-decouple>=3.3

# Data processing
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0
openpyxl>=3.0.0  # Required for reading Excel files with pandas

# Machine Learning
pycaret==3.2.0
scikit-learn<1.3.0  # PyCaret 3.2.0 requires scikit-learn<1.3.0
joblib==1.2.0  # Specific version compatible with PyCaret 3.2.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.25.0

# Logging
loguru==0.7.2

# ASGI Server
uvicorn==0.24.0
